Refactor payments module into submodules
