+++ .gitignore
@@ -29,6 +29,7 @@ fly.toml
 
 # Wasm package
 packages/wasm/breeztech-breez-sdk-spark-v0.1.0.tgz
+.claude/scheduled_tasks.lock
 .claude/settings.local.json
 .claude/worktrees/
 crates/breez-sdk/breez-bench/js/node_modules/
