+++ crates/breez-sdk/core/src/utils/mod.rs
@@ -1,9 +1,9 @@
+pub(crate) mod bitcoin_dust;
 pub(crate) mod contacts_validation;
 pub(crate) mod deposit_chain_syncer;
 pub(crate) mod expiring_cell;
 pub(crate) mod payments;
 pub(crate) mod polling;
-pub(crate) mod send_payment_validation;
 pub(crate) mod token;
 pub(crate) mod utxo_fetcher;
